move is_model_viewable to Utils
